Keep control-plane turn result and stream glue out of the server root
After the turn submit runtime shell moved out, control_plane_server.rs still owned the turn result/stream mapping and SSE glue. This extraction makes control_plane_turn_runtime.rs the authoritative home for turn summary/result mapping and turn stream state helpers so the server root keeps shrinking toward route orchestration.

Constraint: Preserve control-plane turn result and stream behavior exactly while narrowing the host shell
Rejected: Leave result/stream glue in the server root until a larger control-plane rewrite | that would keep the host shell oversized even though the seam already existed
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep control-plane turn mapping and SSE helpers beside the turn runtime module; do not re-embed them into control_plane_server.rs
Tested: ./scripts/cargo-local-toolchain.sh check -p loong; ./scripts/cargo-local-toolchain.sh test -p loong turn_stream_ -- --test-threads=1; ./scripts/cargo-local-toolchain.sh test -p loong turn_result_and_stream_reject_hidden_session_visibility -- --test-threads=1; ./scripts/cargo-local-toolchain.sh clippy -p loong --all-targets -- -D warnings; ./scripts/cargo-local-toolchain.sh fmt --all -- --check
Not-tested: full workspace test suite; non-turn control-plane routes (unchanged in this slice)
